Translation of the Pages Fixes #46
internationalization provided using next/intl library.

content for en and tr languages are in the messages folder.

new route path strategy applied using [locale] param folder.

initial localization is implemented to Categories, Featured courses and GetStarted.
