+++ app/[locale]/Components/FeaturedCourses.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import FeaturedCourse from './FeaturedCourse'
+import { useTranslations } from 'next-intl'
 export default function FeaturedCourses() {
     const mockCourses = [{
         imageUrl: "https://images.unsplash.com/photo-1626908013351-800ddd734b8a?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=MnwyMDUzMDJ8MHwxfHNlYXJjaHwxMXx8bWF4aW1hbGZvY3VzfGVufDF8fHx8MTY1MjIyMDQyNA&ixlib=rb-1.2.1&q=80&w=1080",
@@ -46,10 +47,11 @@ export default function FeaturedCourses() {
         }
     }
     ]
+    const t = useTranslations("Home",)
     return (
         <>
             <h2 className='text-2xl md:text-3xl 2xl:text-4xl font-bold mt-12 mb-4 2xl:mb-12 '>
-                Featured Courses
+                {t("Featured Courses.title")}
             </h2>
             <div className='grid sm:grid-cols-2 2xl:grid-cols-4 gap-2 sm:gap-4 2xl:gap-6'>
                 {mockCourses.map((course, index) => (
