+++ messages/en.json
@@ -3,5 +3,13 @@
         "title": "Discover Passion",
         "description": "Find out what topics you find interesting, learn a new skill & connect with people that are passionate about similar topics.",
         "button": "Get Started"
+    },
+    "Home": {
+        "Featured Courses":{
+            "title":"Featured Courses"
+        },
+        "Categories": {
+            "title":"Categories"
+        }
     }
 }
