+++ messages/tr.json
@@ -3,5 +3,13 @@
         "title": "Tutkunu Keşfet",
         "description": "İlgi duyduğun konuları keşfet, yeni beceriler geliştir ve benzer tutkulara sahip olan insanlarla bağlantı kur.",
         "button": "BAŞLA"
+    },
+    "Home": {
+        "Featured Courses":{
+            "title":"Öne Çıkan Kurslar"
+        },
+        "Categories": {
+            "title":"Kategoriler"
+        }
     }
 }
